Fix broken link in Readme
In the readme this repo covers the different content types that Snyk scans for. Upon trying to open the 'Snyk Container' link I was brought to a dead page as it pointed to an intra-repo path that did not exist. 

Pointed it to the proper location within the snyk docs.